FAC-37: Move GetAssessmentProgressService to assessment package
