+++ src/test/java/org/flickit/flickitassessmentcore/application/service/assessment/GetAssessmentProgressServiceTest.java
@@ -4,7 +4,6 @@ import org.assertj.core.api.Assertions;
 import org.flickit.flickitassessmentcore.application.domain.mother.AssessmentMother;
 import org.flickit.flickitassessmentcore.application.port.in.assessment.GetAssessmentProgressUseCase.Param;
 import org.flickit.flickitassessmentcore.application.port.out.assessment.GetAssessmentProgressPort;
-import org.flickit.flickitassessmentcore.application.service.answer.GetAssessmentProgressService;
 import org.flickit.flickitassessmentcore.application.service.exception.ResourceNotFoundException;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
